subscription and dataloader

+++ graphql-server/src/resolvers/subject.js
@@ -2,10 +2,11 @@ const db = require('../db');
 const { NotFoundError } = require('../errors/errors');
 const { AuthenticationError } = require('apollo-server');
 const pubsub = require('../pubsub');
+const dl = require('../dataloader')();
 const logger = require('pino')({ name: 'create-subject', level: 'debug' });
 
 module.exports.genericSubjectResolver = (rootKey) => {
-  return async (root, args, { dl }, info) => {
+  return async (root, args, context, info) => {
     return dl.subjectById.load(root[rootKey]);
   };
 };
@@ -16,7 +17,7 @@ module.exports.subjectChangedResolver = {
   },
 };
 
-module.exports.getSubject = async (root, args, { dl }, info) => {
+module.exports.getSubject = async (root, args, context, info) => {
   const subject = await dl.subjectById.load(args.id);
 
   if (subject === null) {
@@ -38,12 +39,18 @@ module.exports.createSubject = async (root, args, context, info) => {
     subjectId: subject.id,
   });
 
-  context.pubsub.publish(pubsub.SUBJECT_CHANGED, { subjectChanged: subject });
+  pubsub.pubsub.publish(pubsub.SUBJECT_CHANGED, { subjectChanged: subject });
   return subject;
 };
 
 module.exports.getSubjects = async (root, args, context, info) => {
   const subjects = await db.subject.findAll();
+
+  // Prime dataloader cache
+  for (const subject of subjects) {
+    dl.subjectById.prime(subject.id, subject);
+  }
+
   return subjects;
 };
 
@@ -58,6 +65,10 @@ module.exports.deleteSubject = async (root, args, context, info) => {
     },
   });
 
+  if (subject === null) {
+    throw new NotFoundError();
+  }
+
   await db.subject_history.destroy({
     where: {
       subjectId: { [db.Op.eq]: args.id },
@@ -70,11 +81,12 @@ module.exports.deleteSubject = async (root, args, context, info) => {
     },
   });
 
-  if (subject === null) {
-    throw new NotFoundError();
-  }
-
   await subject.destroy();
+
+  // Clear the dataloader cache
+  dl.subjectById.clear(subject.id);
+  dl.voteBySubjectId.clear(subject.id);
+  dl.voteCountBySubjectId.clear(subject.id);
 
   return true;
 };
@@ -102,26 +114,21 @@ module.exports.updateSubject = async (root, args, context, info) => {
   });
 
   await subject.update(args.input);
+
+  // Clear dataloader cache
+  dl.subjectById.clearAll();
+
   pubsub.pubsub.publish(pubsub.SUBJECT_CHANGED, { subjectChanged: subject });
   return subject;
 };
 
 module.exports.voteCountResolver = async (root, args, context, info) => {
-  const count = await db.vote.count({
-    where: {
-      subjectId: { [db.Op.eq]: root.id },
-    },
-  });
-
-  return count;
+  const result = await dl.voteCountBySubjectId.load(root.id);
+  return result.vote_count;
 };
 
 module.exports.voteResolver = async (root, args, context, info) => {
-  const votes = await db.vote.findAll({
-    where: {
-      subjectId: { [db.Op.eq]: root.id },
-    },
-  });
+  const votes = await dl.voteBySubjectId.load(root.id);
 
   return votes;
 };
